relocate handy function
