+++ apps/web/src/utils/teamUtils.ts
@@ -1,5 +1,11 @@
+import { RoleName } from "@gc-digital-talent/auth";
 import { notEmpty } from "@gc-digital-talent/helpers";
-import { Role, RoleAssignment, UserPublicProfile } from "~/api/generated";
+import {
+  Maybe,
+  Role,
+  RoleAssignment,
+  UserPublicProfile,
+} from "~/api/generated";
 
 export type TeamMember = {
   roles: Array<Role>;
@@ -36,4 +42,29 @@ export const groupRoleAssignmentsByUser = (
   });
 
   return users;
+};
+
+/**
+ * Check to see if user contains one or more roles
+ *
+ * @param checkRoles              Roles to check for
+ * @param userRoleAssignments     Users current role assignments
+ * @returns boolean
+ */
+export const checkRole = (
+  checkRoles: RoleName[] | null,
+  userRoleAssignments: Maybe<RoleAssignment[]>,
+): boolean => {
+  if (!checkRoles) {
+    return true;
+  }
+  const visible = checkRoles.reduce((prev, curr) => {
+    if (userRoleAssignments?.map((a) => a.role?.name)?.includes(curr)) {
+      return true;
+    }
+
+    return prev;
+  }, false);
+
+  return visible;
 };
